@@ -1,6 +1,6 @@
 {
   "name": "reg.angular",
-  "version": "0.4.2",
+  "version": "0.4.3",
   "scripts": {
     "ng": "ng",
     "lint": "eslint . --ext .ts",
@@ -17,30 +17,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "20.1.0",
-    "@angular/cdk": "20.1.0",
-    "@angular/common": "20.1.0",
-    "@angular/compiler": "20.1.0",
-    "@angular/core": "20.1.0",
-    "@angular/forms": "20.1.0",
-    "@angular/material": "20.1.0",
-    "@angular/platform-browser": "20.1.0",
-    "@angular/platform-browser-dynamic": "20.1.0",
-    "@angular/platform-server": "20.1.0",
-    "@angular/router": "20.1.0",
+    "@angular/animations": "20.1.6",
+    "@angular/cdk": "20.1.5",
+    "@angular/common": "20.1.6",
+    "@angular/compiler": "20.1.6",
+    "@angular/core": "20.1.6",
+    "@angular/forms": "20.1.6",
+    "@angular/material": "20.1.5",
+    "@angular/platform-browser": "20.1.6",
+    "@angular/platform-browser-dynamic": "20.1.6",
+    "@angular/platform-server": "20.1.6",
+    "@angular/router": "20.1.6",
     "@types/bootstrap": "5.2.10",
-    "@fortawesome/angular-fontawesome": "2.0.1",
-    "@fortawesome/fontawesome-svg-core": "6.7.2",
-    "@fortawesome/free-solid-svg-icons": "6.7.2",
+    "@fortawesome/angular-fontawesome": "3.0.0",
+    "@fortawesome/fontawesome-svg-core": "7.0.0",
+    "@fortawesome/free-solid-svg-icons": "7.0.0",
     "@ng-bootstrap/ng-bootstrap": "19.0.1",
-    "@ng-select/ng-select": "15.1.3",
-    "@ngx-translate/core": "16.0.4",
-    "@ngx-translate/http-loader": "16.0.1",
+    "@ng-select/ng-select": "20.1.0",
+    "@ngx-translate/core": "17.0.0",
+    "@c": "17.0.0",
     "@popperjs/core": "2.11.8",
     "bootstrap": "5.3.7",
-    "core-js": "3.44.0",
+    "core-js": "3.45.0",
     "jquery": "3.7.1",
-    "ngx-cookie-service": "20.0.1",
+    "ngx-cookie-service": "20.1.0",
     "oidc-client": "1.11.5",
     "run-script-os": "1.1.6",
     "rxjs": "7.8.2",
@@ -48,20 +48,20 @@
     "zone.js": "0.15.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "20.1.0",
-    "@angular/cli": "20.1.0",
-    "@angular/compiler-cli": "20.1.0",
-    "@angular/language-service": "20.1.0",
+    "@angular-devkit/build-angular": "20.1.5",
+    "@angular/cli": "20.1.5",
+    "@angular/compiler-cli": "20.1.6",
+    "@angular/language-service": "20.1.6",
     "@shopify/eslint-plugin": "50.0.0",
     "@types/jasmine": "5.1.8",
     "@types/jasminewd2": "2.0.13",
-    "@types/node": "24.0.13",
-    "@typescript-eslint/eslint-plugin": "8.36.0",
-    "@typescript-eslint/parser": "8.36.0",
-    "eslint": "9.31.0",
-    "eslint-config-prettier": "10.1.5",
-    "eslint-plugin-prettier": "5.5.1",
-    "jasmine-core": "5.8.0",
+    "@types/node": "24.2.1",
+    "@typescript-eslint/eslint-plugin": "8.39.0",
+    "@typescript-eslint/parser": "8.39.0",
+    "eslint": "9.33.0",
+    "eslint-config-prettier": "10.1.8",
+    "eslint-plugin-prettier": "5.5.4",
+    "jasmine-core": "5.9.0",
     "karma": "6.4.4",
     "karma-chrome-launcher": "3.2.0",
     "karma-coverage": "2.2.1",
