@@ -1,5 +1,5 @@
-import { Component } from '@angular/core';
-import { TranslateService } from '@ngx-translate/core';
+import {Component} from '@angular/core';
+import {TranslateService} from '@ngx-translate/core';
 
 @Component({
   selector: 'app-root',
@@ -9,7 +9,7 @@ import { TranslateService } from '@ngx-translate/core';
 export class AppComponent {
   constructor(translate: TranslateService) {
     translate.addLangs(['en', 'hu']);
-    translate.setDefaultLang('en');
+    translate.setFallbackLang('en');
     translate.use('en');
   }
 }
