@@ -1,8 +1,8 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import { HomeComponent } from './modules/home/home.component';
-import { EncounterListComponent } from './modules/encounter/encounter.list.component';
-import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+import {NgModule} from '@angular/core';
+import {Routes, RouterModule} from '@angular/router';
+import {HomeComponent} from './modules/home/home.component';
+import {EncounterListComponent} from './modules/encounter/encounter.list.component';
+import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
 
 export const routes: Routes = [
   {
@@ -26,4 +26,5 @@ export const routes: Routes = [
   imports: [RouterModule.forRoot(routes), NgbModule],
   exports: [RouterModule]
 })
-export class AppRoutingModule {}
+export class AppRoutingModule {
+}
