@@ -1,9 +1,9 @@
-import { Directive, EventEmitter, Input, Output, HostBinding, HostListener } from '@angular/core';
-import { EncounterDetailModel } from '../models/encounter.model';
+import {Directive, EventEmitter, Input, Output, HostBinding, HostListener} from '@angular/core';
+import {EncounterDetailModel} from '../models/encounter.model';
 
 export type SortDirection = 'asc' | 'desc' | '';
 export type SortColumn = keyof EncounterDetailModel | '';
-const rotate: { [key: string]: SortDirection } = { asc: 'desc', desc: '', '': 'asc' };
+const rotate: { [key: string]: SortDirection } = {asc: 'desc', desc: '', '': 'asc'};
 
 export interface SortEvent {
   column: SortColumn;
@@ -22,12 +22,13 @@ export class SortableHeaderDirective {
   @HostBinding('class.asc') get asc() {
     return this.direction === 'asc';
   }
+
   @HostBinding('class.desc') get desc() {
     return this.direction === 'desc';
   }
 
   @HostListener('click') rotate() {
     this.direction = rotate[this.direction];
-    this.sort.emit({ column: this.sortable, direction: this.direction });
+    this.sort.emit({column: this.sortable, direction: this.direction});
   }
 }
