@@ -1,8 +1,8 @@
-import { Injectable } from '@angular/core';
-import { RouterStateSnapshot, TitleStrategy } from '@angular/router';
-import { TranslateService } from '@ngx-translate/core';
-import { Title } from '@angular/platform-browser';
-import { Constant } from '../constants/constant';
+import {Injectable} from '@angular/core';
+import {RouterStateSnapshot, TitleStrategy} from '@angular/router';
+import {TranslateService} from '@ngx-translate/core';
+import {Title} from '@angular/platform-browser';
+import {Constant} from '../constants/constant';
 
 @Injectable()
 export class CustomTitleStrategy extends TitleStrategy {
